[grammar] Remove begin-end expressions and unassigned expressions-as-statements
This change tightens the grammar, mainly to provide better error
messages. It also has the side-effect of breaking some previously-valid
Minispec programs, though they used features that were outside of the
spec.

Main benefit: This change avoids NoViableAlt exceptions on begin-end
statements. These have been a constant source of confusing error
messages. The problem was that a statement could be a begin-end
statement or a begin-end expression followed by a semicolon. A parsing
error was too complex to disambiguate for the parser, which bailed with
a NoViableAlt exception.

There are two changes to achieve this. First, a statement is now
restricted to be exprPrimary followed by ";", rather than any
expression. This disambiguates between begin-end statements and
expressions, case statements and expressions, and prevents things like
unassigned expressions (like "a >> 1;") from being parsed correctly.

Second, since begin-end expressions must end with an expression, this
change somewhat hobbles them. So, since we don't use them and are not
part of the spec, this commit removes begin-end expressions entirely.

I also tried a variant that keeps begin-end blocks without making
parsing harder (we only need to add expression: beginEndBlock), so this
is an option if we want to bring them back.

Finally, note that there is still some room for tightening the grammar,
since there are some exprPrimary cases that are not valid statements.

This change was regression-tested against all lab4/lab5/lab6 commits
from Fall19 and Fall20 so far (~100K tests). This eliminates NoViableAlt
/ "cannot parse" errors on all of these, and the remaining NoViableAlt
errors are pretty easy to interpret.

This change does break 1 intermediate lab4 commit that managed to use
an unassigned expression (bsc disallows most of these), though this was
failing. It does break two lab6 commits from students who managed to use
a begin-end block inside a case expression.