[translate] Custom if-statement elaboration
if statements are now not traversed by the default elaboratorWalker, and
elaborated on exitIfStmt(). When the condition is known at elaboartion
time, only the relevant branch of the if statement is elaborated and
emitted. This avoids executing Integer variable assignments on non-taken
branches. For example,

Integer i;
...
if (i % 2 == 0) i = i + 10;

would previously result in i being incremented even if the if-condition
was false. This is now fixed.

The tradeoff is that this approach will not report some elaboration
errors on non-taken branches. I think this is OK, since we were already
not emitting non-taken branches, so typechecking wasn't done.

An alternative approach would be to have a new type of IntegerContext
level (e.g., DEAD) that held updates to variables below its level. This
would allow elaborating dead code without polluting variables in outer
scopes.